warnings hunt!
cleanup some warnings from IDE
remove some duplication on properties loading
remove useless calls to releaseConnection (http://stackoverflow.com/questions/11449863/resteasy-clientresponse-releaseconnection)


Conflicts:
	src/org/ow2/proactive_grid_cloud_portal/scheduler/client/JobsView.java
	src/org/ow2/proactive_grid_cloud_portal/scheduler/client/SchedulerController.java

+++ src/org/ow2/proactive_grid_cloud_portal/scheduler/client/Job.java
@@ -257,7 +257,7 @@ public class Job implements Serializable, Comparable<Job> {
     }
 
     public int compareTo(Job job) {
-        return ((Integer) this.id).compareTo((Integer) job.getId());
+        return ((Integer) this.id).compareTo(job.getId());
     }
 
     @Override
